refactor(design-tokens): adopt the shared token-control library in the editor extension
